move c4r models to model folder
