+++ Integreat/Integreat.Shared/ViewModels/Resdesign/Extras/Careers4RefugeesViewModel.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Threading.Tasks;
 using Integreat.Shared.Data.Loader;
 using Integreat.Shared.Models;
+using Integreat.Shared.Models.Extras.Careers4Refugees;
 using Integreat.Shared.Services;
 using Integreat.Shared.Services.Tracking;
 using Integreat.Shared.Utilities;
@@ -18,14 +19,14 @@ namespace Integreat.Shared {
         #region Fields
         private INavigator _navigator;
 
-        private ObservableCollection<Careers4RefugeesTemp.CareerOffer> _offers;
+        private ObservableCollection<CareerOffer> _offers;
         private bool _hasNoResults;
         private Func<string, bool, GeneralWebViewPageViewModel> _generalWebViewFactory; // factory generated by AutoFac to resolve a GeneralWebViewPageViewModel instance
 
         #endregion
 
         #region Properties
-        public ObservableCollection<Careers4RefugeesTemp.CareerOffer> Offers {
+        public ObservableCollection<CareerOffer> Offers {
             get {
                 return _offers;
             }
@@ -72,8 +73,8 @@ namespace Integreat.Shared {
             var url = forLocation.Careers4RefugeesUrl;
             try {
                 var offers =
-                    new ObservableCollection<Careers4RefugeesTemp.CareerOffer>(
-                        await XmlWebParser.ParseXmlFromAddressAsync<List<Careers4RefugeesTemp.CareerOffer>>(url, "anzeigen"));
+                    new ObservableCollection<CareerOffer>(
+                        await XmlWebParser.ParseXmlFromAddressAsync<List<CareerOffer>>(url, "anzeigen"));
 
                 // hook up the command
                 foreach (var careerOffer in offers)
@@ -95,7 +96,7 @@ namespace Integreat.Shared {
         private async void OnOfferTapped(object careerOfferObject)
         {
             // try to cast the object, abort if failed
-            var careerOffer = careerOfferObject as Careers4RefugeesTemp.CareerOffer;
+            var careerOffer = careerOfferObject as CareerOffer;
             if (careerOffer == null) return;
             var view = _generalWebViewFactory(careerOffer.Link, false);
             view.Title = "Career4Refugees";
